64_03_06_25_TwoPointer/REST-APP-1/Python_Service/requirements.txt
flask
sentence-transformers
faiss-cpu
PyPDF2
python-dotenv
